Move exec command to the cluster. Call resize2fs for the volumes.
